Fix the symbol lookup issue with symmetric memory __init__
